Rename RenderingStrategy to Rendering
