+++ Adapter/Rendering.php
@@ -4,13 +4,13 @@ namespace Utopia\Detector\Adapter;
 
 use Utopia\Detector\Adapter;
 use Utopia\Detector\Detection\Models\FrameworkType;
-use Utopia\Detector\Detection\RenderingStrategy\Detection as RenderingStrategyDetection;
-use Utopia\Detector\Detection\RenderingStrategy\SSR;
+use Utopia\Detector\Detection\Rendering\Detection as RenderingDetection;
+use Utopia\Detector\Detection\Rendering\SSR;
 
-class RenderingStrategy implements Adapter
+class Rendering implements Adapter
 {
     // add param input for mixed input
-    public function detect(array $inputs, FrameworkType $frameworkType = FrameworkType::NEXTJS): ?RenderingStrategyDetection
+    public function detect(array $inputs, FrameworkType $frameworkType = FrameworkType::NEXTJS): ?RenderingDetection
     {
         // implement logic to detect rendering strategy ie server side, client side etc using framework type
