+++ Detection/Framework/Nextjs.php
@@ -3,6 +3,7 @@
 namespace Utopia\Detector\Detection\Framework;
 
 use Utopia\Detector\Detection\Framework\Detection as FrameworkDetection;
+use Utopia\Detector\Detection\Models\PackageManagerType;
 
 class NextJs extends FrameworkDetection
 {
@@ -13,30 +14,20 @@ class NextJs extends FrameworkDetection
 
     public function getInstallCommand(): string
     {
-        return 'npm install';
+        return match ($this->packageManagerType) {
+            PackageManagerType::NPM => 'npm install',
+            PackageManagerType::YARN => 'yarn install',
+            PackageManagerType::PNPM => 'pnpm install',
+        };
     }
 
     public function getBuildCommand(): string
     {
-        return 'npm run build';
-    }
-
-    public function getInstallCommands(): array
-    {
-        return [
-            'npm' => 'npm install',
-            'yarn' => 'yarn install',
-            'pnpm' => 'pnpm install',
-        ];
-    }
-
-    public function getBuildCommands(): array
-    {
-        return [
-            'npm' => 'npm run build',
-            'yarn' => 'yarn build',
-            'pnpm' => 'pnpm build',
-        ];
+        return match ($this->packageManagerType) {
+            PackageManagerType::NPM => 'npm run build',
+            PackageManagerType::YARN => 'yarn build',
+            PackageManagerType::PNPM => 'pnpm build',
+        };
     }
 
     public function getOutputDirectory(): string
