+++ Detection/Rendering/SSR.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Utopia\Detector\Detection\Rendering;
+
+use Utopia\Detector\Detection\Rendering\Detection as RenderingDetection;
+
+class SSR extends RenderingDetection
+{
+    public function getName(): string
+    {
+        return 'ssr';
+    }
+}
