+++ Detector.php
@@ -4,11 +4,10 @@ namespace Utopia\Detector;
 
 class Detector
 {
-
     private array $inputs;
     private Adapter $adapter;
 
-    public function _construct(array $inputs, Adapter $adapter)
+    public function __construct(array $inputs, Adapter $adapter)
     {
         $this->inputs = $inputs;
         $this->adapter = $adapter;
